requirements.txt
google-adk
numpy
scikit-learn
pandas
flask-cors
flask
uvicorn
pydantic
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery
google-cloud-pubsub
google-cloud-logging
google-cloud-secret-manager
google-cloud-translate
google-cloud-vision
requests
pillow
python-multipart
python-dotenv
asyncio